refactor: cleaning up project
