+++ src/infrastructure/http-clients/maps/map-box/interfaces/forward-geocoding-response.ts
