+++ src/infrastructure/http-clients/maps/map-box/interfaces/index.ts
@@ -1,1 +1,1 @@
-export * from './Imap-box-http-client'
+export * from './forward-geocoding-response'
